second try controller layer

+++ src/main/java/com/example/bigjavahomework/services/implementation/CustomersImplementation.java
@@ -1,11 +1,18 @@
 package com.example.bigjavahomework.services.implementation;
 
+import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
+import com.example.bigjavahomework.resources.CustomersResource;
 import com.example.bigjavahomework.entityes.Customers;
 import com.example.bigjavahomework.repositoryes.CustomersRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import com.example.bigjavahomework.services.CustomersService;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
+
+import static com.example.bigjavahomework.mappers.CustomersMapper.CUSTOMERS_MAPPER;
 
 @Service
 @RequiredArgsConstructor
@@ -13,35 +20,48 @@ public class CustomersImplementation implements CustomersService {
     private final CustomersRepository customersRepository;
 
     @Override
-    public List<Customers> allCustomers() {
-        return customersRepository.findAll();
+    public List<CustomersResource> getAll() {
+        return CUSTOMERS_MAPPER.toCustomersResources(customersRepository.findAll());
     }
 
     @Override
-    public Customers save(Customers customer) {
-        return customersRepository.save(customer);
+    public Optional<CustomersResource> getById(Integer id) {
+        return customersRepository.findById(id).map(CUSTOMERS_MAPPER::toCustomersResource);
     }
 
     @Override
-    public Customers update(Customers customer) {
-        Customers existingCustomer = customersRepository.findById(customer.getId()).orElse(null);
-        if (existingCustomer != null) {
-            existingCustomer.setUsername(customer.getUsername());
-            existingCustomer.setPassword(customer.getPassword());
-            existingCustomer.setName(customer.getName());
-            existingCustomer.setAddress(customer.getAddress());
-            return customersRepository.save(existingCustomer);
+    public CustomersResource create(CustomersResource customersResource) {
+        try {
+            Customers customers = customersRepository.save(CUSTOMERS_MAPPER.fromCustomersResource(customersResource));
+            customersResource.setId(customers.getId());
+            return customersResource;
+        } catch (DataIntegrityViolationException e) {
+            throw new DataIntegrityViolationException("Customer with id " + customersResource.getId() + " already exists.");
         }
-        return customersRepository.save(customer);
+    }
+
+    @Override
+    public CustomersResource update(CustomersResource customersResource, Integer id) {
+        try {
+            Customers customers = customersRepository.getReferenceById(id);
+
+            customers.setUsername(customersResource.getUsername());
+            customers.setPassword(customersResource.getPassword());
+            customers.setName(customersResource.getName());
+            customers.setAddress(customersResource.getAddress());
+
+            return CUSTOMERS_MAPPER.toCustomersResource(customersRepository.save(customers));
+        } catch (EntityNotFoundException e) {
+            throw new EntityNotFoundException("Unable to find customer with id " + id + ".");
+        }
     }
 
     @Override
     public void deleteById(Integer id) {
-        customersRepository.deleteById(id);
-    }
-
-    @Override
-    public Customers findById(Integer id) {
-        return customersRepository.findById(id).orElse(null);
+        if (customersRepository.existsById(id)) {
+            customersRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Unable to find customer with id " + id + ".");
+        }
     }
 }
