+++ src/main/java/com/example/bigjavahomework/services/implementation/ProductsImplementation.java
@@ -1,11 +1,18 @@
 package com.example.bigjavahomework.services.implementation;
 
+import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
+import com.example.bigjavahomework.resources.ProductsResource;
 import com.example.bigjavahomework.entityes.Products;
 import com.example.bigjavahomework.repositoryes.ProductsRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import com.example.bigjavahomework.services.ProductsService;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
+
+import static com.example.bigjavahomework.mappers.ProductsMapper.PRODUCTS_MAPPER;
 
 @Service
 @RequiredArgsConstructor
@@ -13,36 +20,49 @@ public class ProductsImplementation implements ProductsService {
     private final ProductsRepository productsRepository;
 
     @Override
-    public List<Products> allProducts() {
-        return productsRepository.findAll();
+    public List<ProductsResource> getAll() {
+        return PRODUCTS_MAPPER.toProductsResources(productsRepository.findAll());
     }
 
     @Override
-    public Products save(Products product) {
-        return productsRepository.save(product);
+    public Optional<ProductsResource> getById(Integer id) {
+        return productsRepository.findById(id).map(PRODUCTS_MAPPER::toProductsResource);
     }
 
     @Override
-    public Products update(Products product) {
-        Products existingProduct = productsRepository.findById(product.getId()).orElse(null);
-        if (existingProduct != null) {
-            existingProduct.setName(product.getName());
-            existingProduct.setPrice(product.getPrice());
-            existingProduct.setQuantity(product.getQuantity());
-            existingProduct.setWeight(product.getWeight());
-            existingProduct.setInfo(product.getInfo());
-            return productsRepository.save(existingProduct);
+    public ProductsResource create(ProductsResource productsResource) {
+        try {
+            Products products = productsRepository.save(PRODUCTS_MAPPER.fromProductsResource(productsResource));
+            productsResource.setId(products.getId());
+            return productsResource;
+        } catch (DataIntegrityViolationException e) {
+            throw new DataIntegrityViolationException("Product with id " + productsResource.getId() + " already exists.");
         }
-        return productsRepository.save(product);
+    }
+
+    @Override
+    public ProductsResource update(ProductsResource productsResource, Integer id) {
+        try {
+            Products products = productsRepository.getReferenceById(id);
+
+            products.setName(productsResource.getName());
+            products.setPrice(productsResource.getPrice());
+            products.setQuantity(productsResource.getQuantity());
+            products.setWeight(productsResource.getWeight());
+            products.setInfo(productsResource.getInfo());
+
+            return PRODUCTS_MAPPER.toProductsResource(productsRepository.save(products));
+        } catch (EntityNotFoundException e) {
+            throw new EntityNotFoundException("Unable to find product with id " + id + ".");
+        }
     }
 
     @Override
     public void deleteById(Integer id) {
-        productsRepository.deleteById(id);
-    }
-
-    @Override
-    public Products findById(Integer id) {
-        return productsRepository.findById(id).orElse(null);
+        if (productsRepository.existsById(id)) {
+            productsRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Unable to find product with id " + id + ".");
+        }
     }
 }
